foresight-rx/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=6.0
